+++ src/platform_impl/android/mod.rs
@@ -872,7 +872,7 @@ impl Window {
 
     pub fn set_window_icon(&self, _window_icon: Option<crate::icon::Icon>) {}
 
-    pub fn set_ime_position(&self, _position: Position) {}
+    pub fn set_ime_cursor_area(&self, _position: Position, _size: Size) {}
 
     pub fn set_ime_allowed(&self, _allowed: bool) {}