+++ src/platform_impl/macos/window.rs
@@ -1161,10 +1161,11 @@ impl WinitWindow {
     }
 
     #[inline]
-    pub fn set_ime_position(&self, spot: Position) {
+    pub fn set_ime_cursor_area(&self, spot: Position, size: Size) {
         let scale_factor = self.scale_factor();
         let logical_spot = spot.to_logical(scale_factor);
-        util::set_ime_position_sync(self, logical_spot);
+        let size = size.to_logical(scale_factor);
+        util::set_ime_cursor_area_sync(self, logical_spot, size);
     }
 
     #[inline]